+++ fe/src/components/mainPage/issueTable/list/listItems/listItemLeft/ListItemLeft.jsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import styled from 'styled-components';
-import Checkbox from '../../../../../../common/Checkbox';
+import Checkbox from '../../../../../../components/common/checkbox/Checkbox';
 import ListItemLeftContent from './listItemLeftContent/ListItemLeftContent';
 
 const ListItemLeftContainer = styled.div`
